LLL refactor (#2596)
* split out codegen/abi.py into two files

gets rid of circular imports

* fix lint

* change Context vars to be optional

* remove some dead arguments and code

* rename parser.py to module.py

* fix lint

* fix some typos

* remove some dead functions

* remove canonicalize_type and abi_type_of

in favor of properties on the LLL types themselves

* fix a couple incorrect selectors

* fix mypy

* change -f asm to emit space-separated opcodes

instead of line-separated opcodes

also clarify nested asm blocks

* add human-readable labels to asm

* "begin" -> "lll_begin" and fix lint

slightly more descriptive

* remove abi_type_of2

change ABI type to property of the vyper type itself, this reduces
import complexity since abi_types now has very few dependencies.

also move vyper.codegen.abi_types to vyper.abi_types.

* fix a type error
