+++ vyper/builtin_functions/convert.py
@@ -20,12 +20,7 @@ from vyper.exceptions import InvalidLiteral, StructureException, TypeMismatch
 from vyper.utils import DECIMAL_DIVISOR, MemoryPositions, SizeLimits
 
 
-def byte_array_to_num(
-    arg,
-    expr,  # TODO dead argument
-    out_type,
-    offset=32,  # TODO probably dead argument
-):
+def byte_array_to_num(arg, out_type):
     """
     Takes a <32 byte array as input, and outputs a number.
     """
@@ -74,7 +69,7 @@ def to_bool(expr, args, kwargs, context):
                 expr,
             )
         else:
-            num = byte_array_to_num(in_arg, expr, "uint256")
+            num = byte_array_to_num(in_arg, "uint256")
             return LLLnode.from_list(
                 ["iszero", ["iszero", num]], typ=BaseType("bool"), pos=getpos(expr)
             )
@@ -98,7 +93,7 @@ def to_uint8(expr, args, kwargs, context):
             )
         else:
             # uint8 clamp is already applied in byte_array_to_num
-            in_arg = byte_array_to_num(in_arg, expr, "uint8")
+            in_arg = byte_array_to_num(in_arg, "uint8")
 
     else:
         # cast to output type so clamp_basetype works
@@ -167,7 +162,7 @@ def to_int128(expr, args, kwargs, context):
                 f"Cannot convert bytes array of max length {in_arg.typ.maxlen} to int128",
                 expr,
             )
-        return byte_array_to_num(in_arg, expr, "int128")
+        return byte_array_to_num(in_arg, "int128")
 
     elif input_type == "uint256":
         if in_arg.typ.is_literal:
@@ -247,7 +242,7 @@ def to_uint256(expr, args, kwargs, context):
                 f"Cannot convert bytes array of max length {in_arg.typ.maxlen} to uint256",
                 expr,
             )
-        return byte_array_to_num(in_arg, expr, "uint256")
+        return byte_array_to_num(in_arg, "uint256")
 
     else:
         raise InvalidLiteral(f"Invalid input for uint256: {in_arg}", expr)
@@ -315,7 +310,7 @@ def to_int256(expr, args, kwargs, context):
                 f"Cannot convert bytes array of max length {in_arg.typ.maxlen} to int256",
                 expr,
             )
-        return byte_array_to_num(in_arg, expr, "int256")
+        return byte_array_to_num(in_arg, "int256")
 
     else:
         raise InvalidLiteral(f"Invalid input for int256: {in_arg}", expr)
@@ -333,7 +328,7 @@ def to_decimal(expr, args, kwargs, context):
                 expr,
             )
         # use byte_array_to_num(int128) because it is cheaper to clamp int128
-        num = byte_array_to_num(in_arg, expr, "int128")
+        num = byte_array_to_num(in_arg, "int128")
         return LLLnode.from_list(
             ["mul", num, DECIMAL_DIVISOR], typ=BaseType("decimal"), pos=getpos(expr)
         )
