+++ vyper/builtin_functions/functions.py
@@ -4,10 +4,11 @@ import operator
 from decimal import Decimal
 
 from vyper import ast as vy_ast
+from vyper.abi_types import ABI_Tuple
 from vyper.ast.signatures.function_signature import VariableRecord
 from vyper.ast.validation import validate_call_args
 from vyper.builtin_functions.convert import convert
-from vyper.codegen.abi import ABI_Tuple, abi_encode, abi_type_of, abi_type_of2
+from vyper.codegen.abi_encoder import abi_encode
 from vyper.codegen.core import (
     LLLnode,
     add_ofst,
@@ -1870,7 +1871,7 @@ class ABIEncode(_SimpleBuiltinFunction):
         arg_abi_types = []
         for arg in node.args:
             arg_t = get_exact_type_from_node(arg)
-            arg_abi_types.append(abi_type_of2(arg_t))
+            arg_abi_types.append(arg_t.abi_type)
 
         # special case, no tuple
         if len(arg_abi_types) == 1 and not self._ensure_tuple(node):
@@ -1903,7 +1904,7 @@ class ABIEncode(_SimpleBuiltinFunction):
         else:
             encode_input = lll_tuple_from_args(args)
 
-        input_abi_t = abi_type_of(encode_input.typ)
+        input_abi_t = encode_input.typ.abi_type
         maxlen = input_abi_t.size_bound()
         if method_id is not None:
             maxlen += 4
