+++ vyper/codegen/context.py
@@ -16,9 +16,9 @@ class Constancy(enum.Enum):
 class Context:
     def __init__(
         self,
-        vars,
         global_ctx,
         memory_allocator,
+        vars_=None,
         sigs=None,
         forvars=None,
         return_type=None,
@@ -29,32 +29,40 @@ class Context:
         sig=None,
     ):
         # In-memory variables, in the form (name, memory location, type)
-        self.vars = vars or {}
+        self.vars = vars_ or {}
+
         # Global variables, in the form (name, storage location, type)
         self.globals = global_ctx._globals
+
         # ABI objects, in the form {classname: ABI JSON}
         self.sigs = sigs or {"self": {}}
+
         # Variables defined in for loops, e.g. for i in range(6): ...
         self.forvars = forvars or {}
+
         # Return type of the function
         self.return_type = return_type
+
         # Is the function constant?
         self.constancy = constancy
+
         # Whether body is currently in an assert statement
         self.in_assertion = False
+
         # Whether we are currently parsing a range expression
         self.in_range_expr = False
+
         # Is the function payable?
         self.is_payable = is_payable
+
         # List of custom structs that have been defined.
         self.structs = global_ctx._structs
-        # Callback pointer to jump back to, used in internal functions.
-        self.callback_ptr = None
+
         self.is_internal = is_internal
-        # method_id of current function
-        # self.method_id = method_id
+
         # store global context
         self.global_ctx = global_ctx
+
         # full function signature
         self.sig = sig
         # Active scopes
