+++ vyper/codegen/expr.py
@@ -325,7 +325,7 @@ class Expr:
             var = self.context.globals[self.expr.id]
             is_constructor = self.expr.get_ancestor(vy_ast.FunctionDef).get("name") == "__init__"
             if is_constructor:
-                # store memory position for later access in parser.py in the variable record
+                # store memory position for later access in module.py in the variable record
                 memory_loc = self.context.new_variable(f"#immutable_{self.expr.id}", var.typ)
                 self.context.global_ctx._globals[self.expr.id].pos = memory_loc
                 # store the data offset in the variable record as well for accessing
@@ -530,19 +530,19 @@ class Expr:
 
         # If one value of the operation is a literal, we recast it to match the non-literal type.
         # We know this is OK because types were already verified in the actual typechecking pass.
-        # This is a temporary solution to not break parser while we work toward removing types
+        # This is a temporary solution to not break codegen while we work toward removing types
         # altogether at this stage of complition. @iamdefinitelyahuman
         if literals == {True, False} and len(types) > 1 and "decimal" not in types:
             if left.typ.is_literal and SizeLimits.in_bounds(right.typ.typ, left.value):
                 left = LLLnode.from_list(
                     left.value,
-                    typ=BaseType(right.typ.typ, None, is_literal=True),
+                    typ=BaseType(right.typ.typ, is_literal=True),
                     pos=pos,
                 )
             elif right.typ.is_literal and SizeLimits.in_bounds(left.typ.typ, right.value):
                 right = LLLnode.from_list(
                     right.value,
-                    typ=BaseType(left.typ.typ, None, is_literal=True),
+                    typ=BaseType(left.typ.typ, is_literal=True),
                     pos=pos,
                 )
