+++ vyper/codegen/function_definitions/external_function.py
@@ -193,7 +193,7 @@ def generate_lll_for_external_function(code, sig, context, check_nonpayable):
 
     exit = [["label", sig.exit_sequence_label]] + nonreentrant_post
     if sig.is_init_func:
-        pass  # init func has special exit sequence generated by parser.py
+        pass  # init func has special exit sequence generated by module.py
     elif context.return_type is None:
         exit += [["stop"]]
     else:
@@ -205,7 +205,7 @@ def generate_lll_for_external_function(code, sig, context, check_nonpayable):
     func_common_lll = ["seq"] + entrance + body + exit
 
     if sig.is_default_func or sig.is_init_func:
-        # default and init funcs have special entries generated by parser.py
+        # default and init funcs have special entries generated by module.py
         ret = func_common_lll
     else:
         ret = kwarg_handlers
