+++ vyper/codegen/module.py
@@ -1,4 +1,6 @@
-from typing import Any, List, Optional, Tuple, Union
+# a contract.vy -- all functions and constructor
+
+from typing import Any, List, Tuple, Union
 
 from vyper import ast as vy_ast
 from vyper.ast.signatures.function_signature import FunctionSignature
@@ -16,7 +18,6 @@ from vyper.exceptions import (
     StructureException,
 )
 from vyper.semantics.types.function import FunctionVisibility, StateMutability
-from vyper.typing import InterfaceImports
 from vyper.utils import LOADED_LIMITS
 
 # TODO remove this check
@@ -286,17 +287,3 @@ def parse_tree_to_lll(global_ctx: GlobalContext) -> Tuple[LLLnode, LLLnode]:
         runtime = o.copy()
 
     return LLLnode.from_list(o), LLLnode.from_list(runtime)
-
-
-# TODO this function is dead code
-def parse_to_lll(
-    source_code: str, runtime_only: bool = False, interface_codes: Optional[InterfaceImports] = None
-) -> LLLnode:
-    vyper_module = vy_ast.parse_to_ast(source_code)
-    global_ctx = GlobalContext.get_global_context(vyper_module, interface_codes=interface_codes)
-    lll_nodes, lll_runtime = parse_tree_to_lll(global_ctx)
-
-    if runtime_only:
-        return lll_runtime
-    else:
-        return lll_nodes
