+++ vyper/semantics/types/value/array_value.py
@@ -1,6 +1,7 @@
 from typing import Type
 
 from vyper import ast as vy_ast
+from vyper.abi_types import ABI_Bytes, ABI_String, ABIType
 from vyper.exceptions import CompilerPanic, StructureException, UnexpectedValue
 from vyper.semantics import validation
 from vyper.utils import ceil32
@@ -63,10 +64,6 @@ class _ArrayValueDefinition(ValueTypeDefinition):
         # boundary as it's size, instead of giving it a size that is not cleanly divisble by 32
 
         return 32 + ceil32(self.length)
-
-    @property
-    def canonical_abi_type(self) -> str:
-        return self._id.lower()
 
     def set_length(self, length):
         """
@@ -150,9 +147,17 @@ class _ArrayValuePrimitive(BasePrimitive):
 class BytesArrayDefinition(BytesAbstractType, ArrayValueAbstractType, _ArrayValueDefinition):
     _id = "Bytes"
 
+    @property
+    def abi_type(self) -> ABIType:
+        return ABI_Bytes(self.length)
+
 
 class StringDefinition(ArrayValueAbstractType, _ArrayValueDefinition):
     _id = "String"
+
+    @property
+    def abi_type(self) -> ABIType:
+        return ABI_String(self.length)
 
 
 class BytesArrayPrimitive(_ArrayValuePrimitive):
